Add files via upload
The voltage divider is wrongly connected to the Rx pin of the 8266 and pin 11 of the Arduino MEGA. One must disconnect: 1) upper terminal of the 1k resistor from pin 4 of the 8266 and 2) cyan wire from pin 11 of the MEGA and switch them over i.e. the cyan wire goes to the 8266 pin 4 and the upper terminal of the 1k resistor goes to pin 11 of the MEGA. If you leave it as is, no voltage dividing is accomplished and the full 5V level is applied to the Rx pin of the 8266; this can stress the ic to unknown results.